requirements.txt
tensorflow
tensorflow-macos
opencv-python
matplotlib
mlx
keras
tensorflow-metal
pyyaml
scikit-learn
seaborn